THREAT INTELLIGENCE & VULERABILITY
After downloading the run the code the command prompt with the command "npm start"

This project is used to check the vulerability of any file of files and also any type of application.

We can extract what data we want from the database.

I just focussed on CVSS score , TITLE , REMEDY , CONSEQUENCES and description.

Happy Coding :-)!!!